requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
python-docx
langchain-nvidia-ai-endpoints
streamlit
bs4